feat: attempt implementation of signing code
The current code is wrong in some way - I am successfully creating a
file, but when running the verify or parsing step errors occur with the
parsing that is not easy to debug. My next steps are to break up the
signing functions into smaller write functions that can more easily be
tested. This will also require breaking up the signing parsing code.